Export SledStorage as DiskStorage Renamed module to disk
