+++ limitador/src/storage/disk/mod.rs
@@ -3,7 +3,7 @@ use crate::storage::StorageErr;
 mod expiring_value;
 mod sled_storage;
 
-pub use sled_storage::SledStorage;
+pub use sled_storage::SledStorage as DiskStorage;
 
 impl From<sled::Error> for StorageErr {
     fn from(error: sled::Error) -> Self {
